AI-related-project/AI_sound_to_video_on_colab/bin/requirements.txt
protobuf
transformers==4.27.1
cpm_kernels
torch==2.1
torchvision==0.16
gradio
mdtex2html
sentencepiece
accelerate
openai-whisper
gtts